Refactor account month packages
